Fix UNICODE and non-UNICODE builds
If more than a page of locals are used, we cannot link to
__chkstk, so I halved the size of the local buffers when UNICODE.

It might be possible to increase the stacksize, but I do not
know enough about how NSIS works to risk that.

It might be possible to use global variables instead, but that
would prevent this from being run multi-threaded.